requirements.txt
asgiref==3.3.1
coverage==5.5
dj-config-url==0.1.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
dynaconf==3.1.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
